requirements.txt
pyyaml
numpy
torch
onnx
onnxruntime